Backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-genai==1.38.0
python-dotenv==1.0.0
PyYAML==6.0.2
